make best effort to infer unmodeled error message and name (#91)
If incoming error response is unknown error, we only make best effort
to infer the error type. This change puts error response body to error
message so that users have the information they need. This change is
necessary for Node because when error response is unknown, the body
stream will never be consumed, which would use a lot of resources
before stream is flushed.